restructured loaders package
