+++ internal/controller/discovery/loaders/http/http_test.go
