+++ internal/controller/discovery/loaders/loaders.go
@@ -1,4 +1,4 @@
-package discovery
+package loaders
 
 import (
 	"fmt"
